add getPeerInfo in Governance
